feat: Refactor whiteboard state management to use Zustand store; remove legacy useWhiteboardState hook
